requirements.txt
fastapi
uvicorn
python-multipart
Pillow
scikit-learn
numpy
tensorflow
# Note: tensorflow is needed to reconstruct the Keras model from saved weights
